+++ src/main/java/com/charlesluxinger/foodtruck/api/infrastructure/repository/CozinhaRepositoryImpl.java
@@ -23,7 +23,7 @@ public class CozinhaRepositoryImpl implements CozinhaRepository {
 	}
 
 	@Override
-	public Cozinha findById(Integer id) {
+	public Cozinha findById(Long id) {
 		return manager.find(Cozinha.class, id);
 	}
